Support serde_with v1 and v2
* Let the user decide which version to use, or even both.
* Update the CI to test with the powerset of all features.
