+++ Cargo.toml
@@ -20,7 +20,8 @@ iana-time-zone = { version = "^0.1.35", optional = true }
 phf = { version = "^0.10.0", optional = true }
 phf_shared = { version = "^0.10.0", optional = true }
 serde = { version = "^1.0.51", default-features = false, optional = true }
-serde_with = { version = "^1.5.0", default-features = false, optional = true }
+serde_with_v1 = { package = "serde_with", version = "^1.5.0", default-features = false, optional = true, features = ["macros"] }
+serde_with_v2 = { package = "serde_with", version = "^2.0.0", default-features = false, optional = true, features = ["macros"] }
 
 [dev-dependencies]
 serde_json = "=1"
@@ -29,27 +30,29 @@ test-strategy = "=0.2"
 
 [features]
 default = ["by-name", "list", "local"]
-## enables [tz_by_name()] to get a time zone at runtime by name
-##
+
+# Enables [tz_by_name()] to get a time zone at runtime by name.
 by-name = ["phf", "phf_shared", "byte-slice-cast"]
-## enables [TZ_NAMES] to get a list of all shipped time zones
-##
+
+# Enables [TZ_NAMES] to get a list of all shipped time zones.
 list = []
-## enables [local_tz()] to get the system time zone
-##
+
+# Enables [local_tz()] to get the system time zone.
 local = ["by-name", "iana-time-zone"]
-## enables the module [serde_as] to (de)serialize (Utc)DateTimes with [serde][::serde]
-##
-serde-as = ["serde", "serde_with"]
-## make the unparsed, binary tzdata of a time zone available
-##
+
+# Make the unparsed, binary tzdata of a time zone available.
 binary = []
+
+# Enables the module [serde_as] to (de)serialize (Utc)DateTimes with [serde][::serde].
+serde-as = ["serde-as-v1"]
+serde-as-v1 = ["serde", "serde_with_v1"]
+serde-as-v2 = ["serde", "serde_with_v2"]
 
 # Internal feature, used when generating docs. *You* are not supposed to enable this feature!
 docsrs = ["testing"]
 
 # Internal feature, used when testing.
-testing = ["serde-as", "serde/derive", "serde_with/macros", "binary"]
+testing = ["serde?/derive", "serde_with_v1?/macros", "serde_with_v2?/macros"]
 
 [package.metadata.docs.rs]
 features = ["docsrs"]
