+++ src/serde_as/seconds.rs
@@ -2,11 +2,10 @@ use std::fmt;
 
 use serde::de::{Error, Visitor};
 use serde::{Deserializer, Serializer};
-use serde_with::{DeserializeAs, SerializeAs};
 use tz::{DateTime, UtcDateTime};
 
-use super::common::{deserialize_date_time, serialize_date_time};
-use crate::serde_as::common::project_utc;
+use super::common::{deserialize_date_time, project_utc, serialize_date_time};
+use super::serde_with;
 
 /// (De)serialize only the seconds of a (Utc)DateTime as an i64
 ///
@@ -19,23 +18,30 @@ use crate::serde_as::common::project_utc;
 /// to make it [serde] serializable/deserializable.
 ///
 /// ```
+/// # #[cfg(all(feature = "serde-as-v1", not(feature = "serde-as-v2")))]
+/// # pub use ::serde_with_v1 as serde_with;
+/// # #[cfg(feature = "serde-as-v2")]
+/// # pub use ::serde_with_v2 as serde_with;
+/// #
 /// use serde::{Deserialize, Serialize};
 /// use serde_with::serde_as;
 /// use tz::UtcDateTime;
 /// use tzdb::serde_as::Seconds;
 ///
+/// # #[serde_as(crate = "serde_with")]
+/// # /*
 /// #[serde_as]
+/// # */
 /// #[derive(Deserialize, Serialize)]
 /// struct Foo {
 ///     #[serde_as(as = "Seconds")]
 ///     now: UtcDateTime,
 /// }
 /// ```
-#[cfg_attr(feature = "docsrs", doc(cfg(feature = "serde-as")))]
 #[derive(Debug, Clone, Copy, Default)]
 pub struct Seconds;
 
-impl<'de> DeserializeAs<'de, UtcDateTime> for Seconds {
+impl<'de> serde_with::DeserializeAs<'de, UtcDateTime> for Seconds {
     fn deserialize_as<D: Deserializer<'de>>(deserializer: D) -> Result<UtcDateTime, D::Error> {
         struct UnixTpl;
 
@@ -60,13 +66,13 @@ impl<'de> DeserializeAs<'de, UtcDateTime> for Seconds {
     }
 }
 
-impl SerializeAs<UtcDateTime> for Seconds {
+impl serde_with::SerializeAs<UtcDateTime> for Seconds {
     fn serialize_as<S: Serializer>(source: &UtcDateTime, serializer: S) -> Result<S::Ok, S::Error> {
         serializer.serialize_i64(source.unix_time())
     }
 }
 
-impl<'de> DeserializeAs<'de, DateTime> for Seconds {
+impl<'de> serde_with::DeserializeAs<'de, DateTime> for Seconds {
     fn deserialize_as<D: Deserializer<'de>>(deserializer: D) -> Result<DateTime, D::Error> {
         let (secs, tz) = deserialize_date_time(deserializer)?;
         let utc = UtcDateTime::from_timespec(secs, 0).map_err(D::Error::custom)?;
@@ -74,45 +80,8 @@ impl<'de> DeserializeAs<'de, DateTime> for Seconds {
     }
 }
 
-impl SerializeAs<DateTime> for Seconds {
+impl serde_with::SerializeAs<DateTime> for Seconds {
     fn serialize_as<S: Serializer>(source: &DateTime, serializer: S) -> Result<S::Ok, S::Error> {
         serialize_date_time(serializer, source, source.unix_time())
     }
-}
-
-#[cfg(test)]
-#[test]
-fn test_seconds_tuple() {
-    use serde::{Deserialize, Serialize};
-    use serde_json::{from_str, to_string};
-    use serde_with::serde_as;
-
-    use crate::time_zone::europe::BERLIN;
-
-    #[serde_as]
-    #[derive(Debug, Serialize, Deserialize, PartialEq, PartialOrd)]
-    struct UtcStruct(#[serde_as(as = "Seconds")] UtcDateTime);
-
-    #[serde_as]
-    #[derive(Debug, Serialize, Deserialize, PartialEq, PartialOrd)]
-    struct DtStruct(#[serde_as(as = "Seconds")] DateTime);
-
-    #[cfg(not(feature = "testing"))]
-    compile_error!("When testing, use: --features testing");
-
-    let utc = UtcDateTime::new(2022, 3, 1, 15, 20, 37, 0).unwrap();
-    let dt = utc.project(BERLIN).unwrap();
-    assert_eq!(utc.unix_time(), 1_646_148_037);
-    assert_eq!(dt.local_time_type().ut_offset(), 3600);
-
-    assert_eq!(to_string(&UtcStruct(utc)).unwrap(), "1646148037");
-    assert_eq!(from_str::<UtcStruct>("1646148037").unwrap(), UtcStruct(utc));
-    assert_eq!(
-        to_string(&DtStruct(dt)).unwrap(),
-        r#"[1646148037,[3600,false,"CET"]]"#,
-    );
-    assert_eq!(
-        from_str::<DtStruct>(r#"[1646148037,[3600,false,"CET"]]"#).unwrap(),
-        DtStruct(dt),
-    );
 }
